refactor(core-types): rename datasource.ts and primitive-types.ts (#1187)
- datasource.ts → datasourceTypes.ts (P2: types aggregation file)
- primitive-types.ts → primitiveTypes.ts (P2: normalize hyphenated suffix)
- Fix P1 rule: skip role-suffix files regardless of export count

Rollback: git revert <hash> --no-edit
