+++ packages/core-types/src/index.ts
@@ -1,6 +1,6 @@
-export type * from './datasource.js';
+export type * from './datasourceTypes.js';
 export type * from './entity-types.js';
 export type * from './id-types.js';
-export type * from './primitive-types.js';
+export type * from './primitiveTypes.js';
 export type * from './validation-types.js';
 export { toNodeId, toNodeType } from './id-util.js';
